refactor: clean up fuzz executor file
