+++ forge/src/executor/fuzz/strategies/calldata.rs
@@ -1,0 +1,16 @@
+use super::fuzz_param;
+use ethers::{abi::Function, types::Bytes};
+use proptest::prelude::Strategy;
+
+/// Given a function, it returns a proptest strategy which generates valid abi-encoded calldata
+/// for that function's input types.
+pub fn fuzz_calldata(func: &Function) -> impl Strategy<Value = Bytes> + '_ {
+    // We need to compose all the strategies generated for each parameter in all
+    // possible combinations
+    let strats = func.inputs.iter().map(|input| fuzz_param(&input.kind)).collect::<Vec<_>>();
+
+    strats.prop_map(move |tokens| {
+        tracing::trace!(input = ?tokens);
+        func.encode_input(&tokens).unwrap().into()
+    })
+}
